pyproject.toml
[tool.poetry]
name = "streamlit-example"
version = "0.1.0"
description = ""
authors = ["lian.jiang <lian.jiang@opendoor.com>"]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.11.0"
snowflake-connector-python = "^2.7.12"
mysql-connector-python = "^8.0.30"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
